requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
cssselect==1.1.0
cycler==0.10.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
Flask==2.0.1
geographiclib==1.50
geopy==2.1.0
idna==2.10
itsdangerous==2.0.1
jieba3k==0.35.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
munch==2.5.0
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
protobuf==3.17.2
pyparsing==2.4.7
pyproj==3.1.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
sgmllib3k==1.0.0
Shapely==1.7.1
six==1.16.0
soupsieve==2.2.1
stanza==1.2
tinysegmenter==0.3
tldextract==3.1.0
torch==1.8.1
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
